Убрал поиск в шапке аналитики, новое имя файла, исправил замечания
