+++ adaptive_hockey_federation/unloads/mapping.py
@@ -1,0 +1,16 @@
+model_mapping = {
+        "players": ("main", "Player", "players_data.xlsx", "Данные игроков"),
+        "teams": ("main", "Team", "teams_data.xlsx", "Данные команд"),
+        "competitions": (
+            "competitions",
+            "Competition",
+            "competitions_data.xlsx",
+            "Данные соревнований",
+        ),
+        "analytics": (
+            "main",
+            "Player",
+            "analytics_data.xlsx",
+            "Данные аналитики"
+        ),
+    }
